tools/opts: fix missing argument check
Fixes cases such as ngl-player --samples input.mp4. In this case,
opts_parse() is called with ac=2 and ac_max=1 because the code is not
supposed to try to read the last argument; but it does. While ac_max was
properly used for the main loop, it wasn't when trying to read the
argument.