vk/glslang_utils: properly decrease refcount on uninit
If refcount is not 0, then refcount-- won't be either because the
condition is evaluated too late. For some reason GCC and Clang seems to
be able to detect this as an always false condition but do not warn
about it:

    f(int x) { return x && (x-- == 0); }

This is properly compiled to `xor eax,eax; ret` on x86 by both
compilers, but they are not able to emit a single warning.