tests/data: avoid rebuilding the data specs
This weird dance with a spec built using *_count=1 is simply to have
a unique set of type and category. Indeed, the fields with
category="single" may appear several times given an identical type.

Instead, we just build a set, which is much more straightforward.

One side effect of this change is that now arrays and animated buffer in
the block are going to have a length ≠ 1 when testing the single, which
is actually a good thing because it creates more random mis-alignments.