build: move external builddir to their dedicated source dirs
This is useful to make use of the dedicated default build directory
configuration of external projects. For example, sxplayer has a .clangd
file as its root pointing to a relative "builddir" as its default build
dir. Following the project convention allows us, within node.gl, to
benefit from the LSP configuration of sxplayer.