utils/hooks: fix session data cache invalidate
The session info hook could return different data for the same given
session ID currently in cache. One case of this can be reproduced with
the following scenario:

1. spawn a ngl-desktop
2. disable automatic refresh
3. close ngl-desktop
4. change the backend in the ngl-desktop spawning configuration
5. spawn a ngl-desktop again
6. re-enable automatic refresh

At this point, the displayed backend will remain the one from the old
session.

This commit addresses this issue by doing 2 things:

1. update all UI columns when receiving an update event (instead of just
   the status)
2. sending this update event when there is a mismatch between the cached
   session and the data received from the hook