pynodegl: replace the dynamic model with a generated one
This commit adds support for symbols discovery with static analyzers
such as pyright.

Before this commit, the pure Python side of the bindings was created
at runtime using the extreme dynamic properties of the language: at
import time, the node classes and their methods were declared
dynamically and inserted in the globals(). While functional, the
downside of this approach is that introspection tools are unable to
anticipate the structure of the bindings (pyright refuses to import the
module for security reasons).

This means that pynodegl users can not have their editor/IDE/static
analyzers to introspect the bindings and thus are unaware of class
names, their methods, expected arguments, etc. It also means we can not
benefit from a lot of sanity checks offered by these tools.

In this commit, we move from this dynamic model to a generated one,
with the code generated into the pynodegl/__init__.py file during the
execution of the setup.py.