texture_vk: use a dedicated field to determine if the image is wrapped
The wrap parameters will be extended to also support image views and
samplers, thus, we can't rely on a single field (external_storage) to
know which resources we need to destroy in ngli_texture_vk_freep().

This will reduce the diff of the commit adding support for the
Mediacodec interop.