vkutils: remove VK_KHR_deferred_host_operations enums
VK_KHR_deferred_host_operations is provided by the Vulkan headers
starting with version 1.2.162 which is not available on Ubuntu 20.04
(providing only version 1.2.131.2).

Removing the VK_KHR_deferred_host_operations enums (which are not used)
allows the Vulkan backend of node.gl to build on Ubuntu 20.04 (even
though the user still need to install a recent enough version of glslang
as the one provided by Ubuntu 20.04 is outdated).