utils/toolbox/grid: keep transform coords in 2D space
The key change in this commit is AutoGrid.transform_coords()
transforming a 2D coord into a 3D coord now remains in 2D space. This
change has an impact in blending where we need to extend the position to
a 3D space for the Translate.vector parameter.

_equilateral_triangle_coords() also needs to be changed because it is
called to get positions that are directly used for the Translate.vector
parameter (not in _color_positions(), but in
_get_blending_base_objects()). Splitting this change out is challenging
so it is kept in this commit.