buffer: make buffer and layout offset consistent with block view
When the buffer node is a view over a block field, we make sure its
buffer field and its layout offset are mapped to the target buffer.

Currently these fields are ignored because every user (pass and
geometry) manually read the content of the block when the buffer is
a block field viewer. This will be addressed in the following commits.