buffer: remove now unused block_field from buffer_info
The block node pointer is still required in the ngli_node_buffer_*()
functions. We can not remove it from there and use opts->block instead
because the block option is only present in some of the node buffers
but not in all of them (Buffer* has it, but AnimatedBuffer* for instance
has a different options set).