glsl: ensure color is always positive before calling pow() in sRGB conversions
From the pow() documentation:

    Results are undefined if x<0 or if x=0 and y=0.

Fixes undefined behaviour on multiple GPUs/backends.

If the color is negative the branches with pow() are not supposed to be
taken, but they are calculated anyway and end up being used in the mix
code: mix(v, NaN, 0.0). Even if 0% of NaN is taken, it is part of the
formula so the undefined behaviour is propagated.

We decide to clamp only the values used by pow() and not the input color
in order to preserve the formula for negative colors: indeed, it might
make sense to temporarily have negative color components (preserving the
proportions between them) and have them corrected up later down the
pipeline. Clamping only the values used by pow() has no impact over the
expected result because the branch is entirely discarded.

The CPU functions ngli_colorconv_srgb2linear() and
ngli_colorconv_linear2srgb() are not affected because the pow() branches
are never taken with negative values.