build: fix glslang version check
The glslang version check is currently broken if the detected version
matches the minimum version. For instance, if the detected version is
11.1.0, glslang_ver is set to '11.1' and compared against '11.1.0'
using the version_compare method which evaluates to false.

This commits fixes this issue by taking into account the patch level of
glslang (which corresponds to the minor version) when building the
glslang_ver variable.