utils/assets: replace H264 mire with a cat video
The H264 mire is in YUV444P. This typically causes the VideoToolbox
decoder to fallback on its own software decoder which fails to decode
the stream. Unfortunately, it fails too late for a fallback on a working
software decoder (because the Apple decoder accepts the stream but fails
to return frames).

In this commit we replace that video with a 169K H264/YUV420P video of a
cat (of about 10 seconds).

The original mire is kept for future experiments, but it is disabled
from the available samples in all scenes.