block: remove ref/unref mechanism
This is the first commit of a serie that will respectively inline the
following functions (ordered by call order):
- ngli_node_block_ref
- ngli_node_block_init
- ngli_node_block_upload
- ngli_node_block_unref
…into:
- block_init
- block_prepare
- block_update
- block_uninit

For the ref/unref case, since nodes are already ref-counted this
duplicated mechanic can be merged into init/uninit.

In node_buffer.c, the explicit ngli_node_block_{ref,unref} calls go away
because the internal nodes API is handling the init/uninit calls itself.

Similarly in pass, we know the resources (in this case blocks) will be
initialized before Render/Compute.

We note that after this change ngli_node_block_get_gpu_size() is
identical to ngli_node_block_get_cpu_size(). This makes no difference
with the current state because in practice the refcount is already
always > 0 for the whole scene lifetime (ref and unref are called from
init and uninit callbacks from other nodes already). The separate
function is still kept because we might have a prefetch/release
mechanism in the future.