block: replace ngli_node_block_init() with block_prepare()
Since this is now a prepare callback, it is responsible for forwarding
the call down to the children.

This re-entrant function was initially called from ngli_pass_prepare().
We can safely remove that call after moving to the block prepare
callback because all pass users (Render and Compute) explicitly call
ngli_node_prepare_children() on themselves. This causes the Block
prepare callback to be called pretty much at the same time as before
this commit.

Similarly, the block init call in ngli_node_buffer_init() is not needed
anymore because the prepare callbacks automatically cascade and reach
the underlying block prepare.