buffer: fix memleak with count based buffers
Filename and count based buffers do allocate an heap buffer. Block and
Data on the other hand are respectively a view over another buffer and
a buffer based on the option parameter already allocated.

We can not write the condition in the form "o->filename || o->count"
because count may be used in other scenarios (typically to select a
slice of any given buffer).