renderother: add support for drawing extra resources
This is similar to the resources drawing support recently added in pass,
but in a simplified way. It will be used in the next commits.

The reset is done unconditionally in the uninit but not initialized in
the init because the init is called too late: each render other that
wants to draw resources has to initialize it explicitly.