hwmap_vaapi_{gl,vk}: sync surface before using it
vaSyncSurface() ensures that all pending operations on the render target
have been completed.

This fixes visual glitches when seeking with the VAAPI/Vulkan
acceleration. This notably fixes the media_queue test.

While the OpenGL backend does not seem to be affected (no visual glitch
appears), we also call vaSyncSurface() there to make it consistent with
the Vulkan backend.