buffer: replace ngli_node_buffer_init() with explicit prepare callbacks
The rationale and explanations of this change are pretty much the same
as with the move from ngli_node_block_init() to block_prepare() in a
previous commit.

The prepare callback in AnimatedBuffer* and StreamedBuffer* do not
require the block code so the functions are simpler than
buffer_prepare(). We also do not bother keeping the assert since the
buffer is allocated unconditionally in the init function (this contrasts
with the Buffer node that has a more complex logic with its Block view
system).

Also, they do not have the explicit call to ngli_buffer_upload() because
contrary to Buffer* nodes, they are dynamic nodes, so the upload will be
called through ngli_node_buffer_upload().

A large chunk in the Geometry node goes away because it was only
responsible for calling ngli_node_buffer_init(), but this is now handled
by the prepare callback cascading.