buffer: remove ref/unref mechanism
Similarly to the changes happening in block, this is the first commit of
a serie that will respectively inline the following functions (ordered
by call order):
- ngli_node_buffer_ref
- ngli_node_buffer_init
- ngli_node_buffer_upload
- ngli_node_buffer_unref
…into:
- buffer_init
- buffer_prepare
- buffer_update
- buffer_uninit

One of the difference with the Block node is that we have multiple
buffer node implementations: Buffer*, AnimatedBuffer* and
StreamedBuffer*. With this commit each buffer node implementation
becomes responsible for explicitly creating and destroying its buffer
object.

We deal with the block_info.block case only in the Buffer* nodes because
AnimatedBuffer* and StreamedBuffer* are not using it.

Geometry and pass do not need to keep track of their buffer nodes
anymore because the internal node ref counting system takes care of it.