nodes: ensure nodes are updated after their activation
Fixes a crash with the following graph, when drawing sucessively at 10.0,
0.0 and 10.0:

                .-------------(Group 1)------------.
                |                                  |
                v                                  v
       (TimeRangeFilters 1) [cont 0]      (TimeRangeFilters 2) [noop  -1]
                |           [noop 1]               |           [cont   9]
                |                                  v           [noop  11]
                |                               (Group 2)
                |                                  |
                '-----------------.----------------'
                                  |
                                  v
                         (TimeRangeFilters 3) [cont 0]
                                  |           [noop 1]
                                  v
                  .------------(Group 3)---------.
                  |                              |
                  v                              v
            .---(RTT)---.                   (RenderTexture)
            |           |                        |
            |           |                        |
            v           v                        |
        (Identity)  (Texture)<-------------------'

Before this commit, the last draw call (at 10.0) produces a crash:

1. first draw call (at 10.0) disables the left branch and enables the
   right one. The right branch is then updated/drawn (Group 1,
   TimeRangeFilters 2, Group 2 and TimeRangeFilters 3)
2. second draw call (at 0.0) enables the left branch and disables the
   right one. The left branch is then updated/drawn (Group 1,
   TimeRangeFilters 1, TimeRangeFilters 3, Group 3, RTT, RenderTexture).
   At this point, Group 2 is disabled and has its last_update_time still
   set to 10.0. TimeRangeFilters 3 is allowing the descent into the
   graph.
3. third draw call (at 10.0), disables the left branch (and de-allocates
   its resources) and enables the right one. However, during the update
   operation, the descent is blocked by Group 2 which has its
   last_update_time equal to the current update time. This, in turn,
   prevents TimeRangeFilters 3 from being updated leading to the node
   keeping its previous state, allowing the descent into the graph while
   it should not for the current draw time. This ultimately leads to a
   crash because the draw operation is called on a branch that has been
   disabled/de-allocated.

This commit ensures that the nodes are updated after their activation.
In the above scenario, it allows Group 2 to be updated when activated
which, in turn, allows TimeRangeFilters 3 to be updated as it should be,
fixing the crash.

The reason why Group 2 has its last_update_time still set to 10.0 at
step 2 is because node_release() has not been called on it (which does
reset the last_update_time). This is due to a regression introduced in
d83b9e49f5ffcdb2d3f89cef0e1aa38a470c2dab¹ reducing the number of nodes
onto which we call node_prefetch()/node_release().

In order to avoid re-inserting a large number of nodes in the activity
check array, we instead add a special condition detecting that
particular situation.

There were multiple other alternative ways of fixing the issue, but this
one is probably the simplest and most optimal. Typically the reason we
reset the time at prefetch instead of at release like before is because
detecting a release is much harder as it can be done only when
everything has been visited: only one use is enough to imply a prefetch,
but a release is conditioned by the lack of any use (it requires full
knowledge).

Thanks to Quentin Despinoy <qdespinoy@gopro.com> for helping me tracking
down this issue.

[1]: this commit is present in release v2022.4, v2022.5 and v2022.6

Signed-off-by: Clément Bœsch <u@pkh.me>