utils/misc: always define a medias list
SceneCfg is going to become a dataclass where we will have to type the
medias list. If we allow a NoneType, many false warnings will be raised
where we dereference medias unconditionally.

A user building a SceneCfg is still able to force an empty list for the
medias (typically to prevent the ngl-control hooks from uploading random
files to a remote device), but using [] instead of None.