api,params: change data size to size_t
Data type parameters are used in 2 node types: AnimKeyFrameBuffer and
all the Buffer* nodes.

AnimKeyFrameBuffer is using animkeyframe_opts to store the parameter,
which is shared amongst animation routines. AnimatedBuffer reads the
data field and transfers it to a buffer_info, so this struct needed an
update of its data_size as well.

Buffer* use a local buffer_opts, and similary to AnimatedBuffer
transfers the data field into a buffer_info.