pipeline_gl: also insert memory barriers before executing the pipeline
This ensures that all memory accesses initiated prior to this barrier
complete. More specifically, this fixes writes after reads hazards.

Currently, a barrier is automatically inserted after a pipeline if
the pipeline is doing incoherent memory accesses (writes to Images or
SSBO). This covers the read-after-write cases but not the write-after-
read cases.

For example, without this patch, the following scenario can occur:

[Draw SSBO (ro)], [Compute SSBO (w)], [Barrier]

A barrier is missing between the Draw and the Compute pipelines,
potentially causing visual glitches as the Compute pipeline can start
before the Draw pipeline completes. With this patch, a barrier is
inserted automatically before and after pipelines performing incoherent
memory accesses. This fixes the above scenario which now looks like this:

[Draw SSBO (ro)], [Barrier], [Compute SSBO (w)], [Barrier]

See, the OpenGL documentation¹ regarding barriers:

 SHADER_STORAGE_BARRIER_BIT: Memory accesses using shader buffer
 variables issued after the barrier will reflect data written by shaders
 prior to the barrier. Additionally, assignments to and atomic operations
 performed on shader buffer variables after the barrier will not execute
 until all memory accesses initiated prior to the barrier complete.

[1]: https://registry.khronos.org/OpenGL/specs/gl/glspec46.core.pdf