tools/desktop: move scene construction earlier
This is going to be needed because the aspect used in get_viewport will
be extracted from the scene. It is placed consistently with other tools
where get_viewport() is also called after building the scene.

Note that the scene unref is moved in the exit path to handle the new
error code paths.