api: remove TimeRangeMode* nodes
This change essentially simplifies the usage of the TimeRangeFilter.

Working in key-frames is unintuitive and unnecessarily complex, so we
instead use start and end absolute time markers.

Having multiple ranges is also rare and a source of complexity, so the
node now only manages one single time segment. If multiple time segments
are required, one can simply use a Group with multiple TimeRangeFilter.
A side effect of having a single time segment is that the max_idle_time
is not necessary anymore: this option only worked between ranges within
a single TimeRangeFilter. While not exactly equivalent, an alternative
is to simply increase the prefetch time of the scenes but the first one.