eval: add mod_e(), mod_f() and mod_t() modulo functions
The euclidean version is argued to be the most appropriate from a
mathematical point of view, while the floored and truncated version are
commonly found in various languages.

The ceil version is left out since it appears that almost no language
implement it, and the round version is also left out as the rounding
could imply all kind of different rounding.

See https://en.wikipedia.org/wiki/Modulo for more information on the
different variants.