pgcraft: simplify textures and images binding logic
Simplifies the textures and images binding logic by letting pgcraft
always decide the binding points even if the context does not support
explicit bindings. On such context, we execute a post processing pass
on the program to enforce the bindings according to the pgcraft
resources description.

This also greatly simplifies the OpenGL pipeline code.