text: move 2D canvas logic in a dedicated text API
This new text unit is responsible for defining a 2D canvas (of size
(text->{width,height} and placing characters in that space, honoring
various settings (currently only padding, but will be extended).

In the text node on the other hand, we are still handling the mapping of
these rasterized characters in the 3D space.

For simplicity reasons the font rasterization is currently still managed
in the text node and drawutils modules, but they will slowly be
transfered into that new text module.