text: split pipeline subdesc into two distinct pipelines (fg and bg)
It's a coincidence that both pipelines have identical inputs (color and
opacity), and it will not hold true in the future. Thus, we split the
structure into 2 distinct ones (and share the common part).

We consider the modelview and projection matrix to be common to both,
but color and opacity to be specific.