text: allow drivers to attach tags to returned characters
These are simple combinable semantic tags that can be used to intepret
the array of characters into distinct chunks.

This introduce the possibility of characters not being glyph, which we
currently ignore.

We use a tag mechanism because we could very well have visible
linebreaks such as "↲" depending on what the text engine decides,
meaning these characters would have both a glyph and a tag. Currently,
the builtin text driver doesn't exploit these features.