text: switch to instancing
While this change is likely less efficient compared to a single
geometry, it offers much more flexibility with regards to transmitting
character-specific data down to the shaders.

Both vertices and uvcoords were previously containing 4 information each
(one for each edge). Now the data is a unique information for each
character. This implied a change in the nature of the data itself.

The 4 vec3 vertices are replaced with a transform matrix that's being
applied a on a constant normalized coordinate in the vertex shader to
obtain the actual vertex position.

Similarly, the 4 vec2 uvcoords are replaced with an atlas identifier,
which the vertex shader itself uses to locate the position in the atlas.

Along these changes, you may notice that the padding in the atlas goes
away and is replaced with an appropriate clamping in the fragment
shader.

Also, the indices goes away because we're now hardcoding the UVs in the
vertex shader.

Splitting this commit in multiple distinct changes was unfortunately way
too challenging because all of them imply quite convoluted intermediate
states.