internal: remove legacy uniforms support
The related GL functions are kept for internal use (notably for texture
binding).

While it greatly simplifies the graphics abstraction layer, using
UBOs over legacy uniforms without explicit synchronization and double
buffering incurs a performance penalty. This is something that will be
addressed later on by introducing explicit sync mechanisms and double
buffering that benefits all the backends.

The performance penalty on an AMD RX6600 rendering the benchmark scene
in 320x240 (with computes) is about 5-10% (going from an average of 2000
fps to 1800 fps for the worst run).