text: rename atlas_coords IO variable to coords
In the fragment shader the coordinates can only be related to the
texture picking. This commit is preparing the following commits where
the atlas coordinates will be directly fed to the vertex shader and we
will need a different name. In the context of the vertex shader we will
preserve the naming "atlas_coords" for the input because we want to
avoid the confusion with the geometry coordinates.