internal: revamp the pipeline vertex buffers API
Prior to this commit, the pipeline API worked directly with vertex
attributes each pointing to a separate buffer. This made the API
tedious to use, in particular with mat3 and mat4, as one needed to call
ngli_pipeline_compat_update_attribute() for every attribute making up
the matrix.

This commit makes the pipeline API works with vertex buffers with each
vertex buffer referencing at least one attribute, making the use of
interlaced vertex data easier.

The vertex buffers/attributes state is also moved from the
pipeline_layout structure to the pipeline_graphics structure. This
separation between the vertex_state structure and the pipeline_layout
structure maps better to the modern APIs.

On the pgcraft side, the interface for declaring attributes remains
unchanged. However, it will be reworked later on to use the same
model as the pipeline, allowing the user to declare interlaced vertex
attributes.