rendertarget_vk: rework and fix render pass creation
If pResolveAttachments is not NULL, each of its elements
corresponds to a color attachment at the same index in
pColorAttachments, thus, pResolveAttachments must have the same
element count as pColorAttachments. If a color attachment is not
resolved, its corresponding entry in pResolveAttachments must use
VK_ATTACHMENT_UNUSED. See the Vulkan documentation¹.

Before this commit, the code responsible for creating render passes was
not following that logic. In practice, it is not an issue as nope.gl
currently configures the rendertargets with either no resolve attachment
or a resolve attachment per color attachment.

[1]: https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkSubpassDescription.html