vk: use common ngli graphics calls where appropriate
In the upcoming commits, common ngli graphics calls will handle more
non-backend specific logic, for example: initializing common fields,
validating parameters and states.

The downside of this approach is that we currently lose some error
information in the gpu_ctx_vk and texture_vk modules. In practice, this
is not an issue for the gpu_ctx_vk module as the error is ultimately
converted to the corresponding ngl ret code (without any logging). For
the texture_vk module and the Vulkan codebase in general, we need to
figure out a proper way to report Vulkan errors.