text: fix texture update misses
When the mutable atlas flag is set, the texture needs to be updated
anytime ngli_text_set_string() is called. Before this commit, we were
updating it only when the number of characters grew, which is not enough
because the texture changes systematically.

A return code check is also added to the function to make sure it's
working as expected.