ci: wait for Vulkan installers to finish before proceeding further
The installers are GUI based meaning we might get control back before
the install ends. This causes the vulkan DLL to not be available during
configure, causing all kind of runtime linking issues.

Piping will wait for the program to terminate, so we're using a noop
piping here.

Debugged and analyzed with Matthieu Bouron.

Signed-off-by: Matthieu Bouron <matthieu.bouron@gmail.com>