tests: fix error code in case ngl-test is called without reference file
sys.exit(data) will print the data before exiting, but the error code
won't be zero if the data actually contains something.

There are two main scenarios where ngl-test is called without a
reference file:
1. for test functions with no reference (for example tests/api.py)
2. manual call from a developer

The first case in unaffected: our tests do asserts in case there is an
error, they are not relying on a non nul return value.

On the other hand this commit prevents the 2nd case from exiting with an
error, and still prints what the function returns.