utils/com: allow specifying a simple module instead of a package
Currently we allow:
- a "package module", such as `pynopegl_utils.examples`
- a script pointing to a module, such as
  `pynopegl-utils/pynopegl_utils/examples/misc.py`

This commit allows to specify `pynopegl_utils.examples.misc` as well.

The presence of ".py" at the end of the string was used to distinguish
between the 2 cases, but that doesn't work anymore: the string itself is
not enough to differenciate a module from a package. So now the string
contains the complete import, which is why we then have to reduce it to
a simpler form in the UI (rsplit chunks).