texture: change default filtering from nearest to linear
From a user perspective, nearest is much less common, in particular in
the context of compositing with multimedia inputs. This saves the hassle
for the users to specify "linear" for both filters most of the time, and
it also prevents a less than desirable output when forgotten.

In the examples scenes, all the textures with no filtering specified are
left unchanged (and thus get switched from nearest to linear) with the
following exceptions:
- the audio texture in `misc.audiotex`
- the depth texture in `misc.cube`
- the noise texture in `misc.mountains`
- the buffer texture in `transforms.animated_buffer`

`misc.lut3d` and `misc.bufferdove` also have their filtering logic
slightly adjusted to take into accouunt the user input.

In the tests we explicit "nearest" to keep the references as is, with
the exception of `api_media_sharing_failure` which is left unchanged
because the visual output doesn't hand an impact on the success of the
test.