utils/qml/ngl_widget: remove livectls check
ngl.get_livectls() always returns a (potentially empty) mapping. If
scene is None the function crashes, so this check is only used meant to
check if there are items. Since the following for-loop works the same if
the dict is empty, we can safely remove this check.

There is one functional change though: now a signal is emitted with a
new empty model. This was actually missing previously and is more
correct.

This diff is best observed without whitespace changes.