utils/config: make path management more friendly with Windows
os.path.expanduser("~") works on Windows, but it does not prefix the
path with the drive. While this does not seem to be an issue using
Path.home() looks more correct.

The generated paths now use proper separators and follow the posix
convention even on Windows (to make it consistent with the scripts
path).