utils/viewer: implement export on Windows
The subprocess API does not curently support the `pass_fds` argument on
Windows. Here is an overview of the workaround proposed in this commit.

File descriptors are non-inheritable on Windows (os.set_inheritable()
does not work), however, the handles that back them are. We take
advantage of the MSVC runtime to get the file handle that backs the pipe
file descriptor and make it inheritable.

Then, we pass the pipe file handle to the FFmpeg command instead of the
file descriptor and we instruct the subprocess command to not close
the fds of our current process so it does not close the pipe fd (and
its underlying file handle). We also instruct subprocess to execute a
dedicated python module that converts the specified file handle back to
a valid fd and call FFmpeg with that fd.