ngl-control: remove ability to control the medias list as a user
This gets in the way of the incoming code refactoring. If we want to
customize the medias of a given scene, this should be handled with a
scene build parameters (or editing the scene code).

This commit only removes the overload of default medias, it doesn't
affect the file upload by the hooks.