utils/viewer: align build parameters and live controls within a grid
The instantiator creates QtObjects containing 2 items, one for each
column of the grid. To store them, we use the properties, as QtObjects
do not have placeholders such as data, children or resources like Items
do. Then for each of these 2 items, we assign them to the row and column
in the grid.

Alternatively, we could use an Item and resources (invisible children),
but a QtObject is more lightweight.

This commit diff is best observed without whitespace changes.

This change was made possible thanks to the help from grecko (@oKcerG on
GitHub) from the Qt discord server.