utils/qml: simplify further the creation of scene controls
The SceneWidget wraps the whole grid fitting logic so that each UI
elements only have to declare their custom part (the widget itself).

The window had to be passed down because the parent chain seems to be
constructed too late for the dialogs (color and file) to be able to find
back the root window. So instead we're being explicit about it.

This change was again made possible thanks to Pierre-Yves Siret
(@oKcerG).