tests/shape: make shape_precision_iovar use ngl_uvcoord instead of ngl_out_pos
We cannot rely on ngl_pos_out as it is not backend agnostic, clip space and NDC
space are not the same across backends. We could have used ngl_position here,
but since ngl_uvcoord coordinates are already in the [0;1] range, we choose the
latter.

This commit makes the test have the same output for both onscreen and offscreen
scenarios with the OpenGL backend.