scene: prevent ngl_scene root replacement for configured graph
In the incoming commits, the graph is going to be altered when
associated with a scene and restored to its original state when
detached. Since this graph modification state will be relied upon in the
rendering, it means we won't be able to detach the graph from the scene
until the rendering ended.