nodes: prevent graph structure changes after scene association
These operations are undefined behaviour since the introduction of the
scene pointer in the commit "check for node sharing within the scene
context".

A live change test had to be simplified because it is not possible to
change the structure of the scene anymore, even after being detached
from the context.