tests: force a consistent aspect ratio and resolution
Currently the rendering context for the tests does not have a viewport.
If we were to honor it, many tests would end up with black bars, which
which we likely do not want to include in the fingerprints and cue
points picking.

This commit adds an assert to make sure the assumption is always met. We
may remove it (maybe conditionally) in the future in case we want to
test letterboxing and other similar mismatching configuration.

The test references are updated due to the scaling going away (the GL
scaling may differ from the one used by PIL to downscale the image) and
various resolutions being adjusted to saner values.