ngpu/gl: drop support for OpenGL < 4.5 and OpenGLES < 3.1
OpenGL 4.5 was released on August 11, 2014, and OpenGL ES 3.1 was released in
March 2014.

Desktop GPUs from the past 4-5 years support OpenGL 4.5 or higher, and mobile
SoCs from the same period support OpenGL ES 3.1 or higher. Moreover, according
to the Android Dashboard, OpenGL ES 3.1 or higher represents more than 92% of
Android devices.

The biggest downside of this change is that nope.gl will no longer support
OpenGL on macOS or iOS, which will break the ngl-viewer and ngl-diff tools as
they rely solely on OpenGL. The default backend for these platforms will be
switched to Vulkan (MoltenVK).

While this change simplifies the code, it will also ease its maintenance and
enable the use of "modern" features in more code paths.